+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+-   `customCssClasses` to `Form` class
+
+### Added
+
 -   `emailAttachmentsEndpoint` to `FormSubmissionEventConfiguration` class
 -   `UploadEmailAttachment` to `FormsClient` class
 -   `WorkflowAttachmentUploadCredentials` to models